+++ requests_auth/__init__.py
@@ -9,6 +9,7 @@ from requests_auth._oauth2.common import OAuth2
 from requests_auth._oauth2.authorization_code import (
     OAuth2AuthorizationCode,
     OktaAuthorizationCode,
+    WakaTimeAuthorizationCode,
 )
 from requests_auth._oauth2.authorization_code_pkce import (
     OAuth2AuthorizationCodePKCE,
@@ -55,6 +56,7 @@ __all__ = [
     "AzureActiveDirectoryImplicitIdToken",
     "OAuth2AuthorizationCode",
     "OktaAuthorizationCode",
+    "WakaTimeAuthorizationCode",
     "OAuth2ClientCredentials",
     "OktaClientCredentials",
     "OAuth2ResourceOwnerPasswordCredentials",